% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acpr}
\alias{acpr}
\title{\code{sf} Multipolygon data for the Aged Care Planning Regions boundaries for Australia (2018 edition).}
\format{
\subsection{\code{acpr}}{

A \code{sf} object.

shows the boundaries of the Aged Care Planning Regions in Australia (2018 edition).
}
}
\source{
\url{https://www.gen-agedcaredata.gov.au/resources/access-data/2020/january/aged-care-planning-region-maps}
}
\usage{
acpr
}
\description{
\code{sf} Multipolygon data for the Aged Care Planning Regions boundaries for Australia (2018 edition).
}
\keyword{datasets}
