% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhn}
\alias{lhn}
\title{\code{sf} Multipolygon data for the Local Hospital Network boundaries for Australia.}
\format{
\subsection{\code{lhn}}{

A \code{sf} object.

shows the boundaries of the Local Hospital Network in Australia.
}
}
\source{
\url{https://hub.arcgis.com/datasets/ACSQHC::local-hospital-networks/explore}
for most of Australia but \url{https://discover.data.vic.gov.au/dataset/dhhs-service-areas}
for Victoria, \url{https://catalogue.data.wa.gov.au/dataset/health-services-health-006}
for Western Australia, and \url{https://data.sa.gov.au/data/dataset/local-health-networks}
for South Australia.
}
\usage{
lhn
}
\description{
\code{sf} Multipolygon data for the Local Hospital Network boundaries for Australia.
}
\keyword{datasets}
