% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb21_poly_A}
\alias{mb21_poly_A}
\title{\code{sf} Polygons data for the meshblocks (2021 edition) for Australia including population counts (\code{Person}). (NSW and VIC only.)}
\format{
\subsection{\code{mb21_poly_A}}{

A \code{sf} object.
}
}
\source{
Mesh Blocks 2021 shapefile \url{https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files} and population data \url{https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/latest-release}.
}
\usage{
mb21_poly_A
}
\description{
\code{sf} Polygons data for the meshblocks (2021 edition) for Australia including population counts (\code{Person}). (NSW and VIC only.)
}
\keyword{datasets}
