% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmm19}
\alias{mmm19}
\title{A \code{data.frame} for the SA1 (2016 edition) and associated Modified Monash Model (MMM) classification (2019 edition) for Australia including population counts (\code{Person}). (States othe than NSW and VIC only.)}
\format{
\subsection{\code{mmm19}}{

A \code{data.frame}.
}
}
\source{
Mesh Blocks 2021 shapefile \url{https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files} and population data \url{https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/latest-release}.
}
\usage{
mmm19
}
\description{
A \code{data.frame} for the SA1 (2016 edition) and associated Modified Monash Model (MMM) classification (2019 edition) for Australia including population counts (\code{Person}). (States othe than NSW and VIC only.)
}
\keyword{datasets}
